fixing signed with unsigned comparism

+++ exercises/levenshtein-distance/example.cpp
@@ -8,19 +8,19 @@ int levenshtein::distance(const std::string& source, const std::string& target)
 	std::vector<std::vector<int>> table(source.size() + 1
 		, std::vector<int>(target.size() + 1, 0));
 	
-	for (int i = 0; i<source.size() + 1; i++) 
+	for (size_t i = 0; i<source.size() + 1; i++) 
 	{
 		table[i][0] = i;
 	}
 
-	for (int j = 0; j<target.size() + 1; j++) 
+	for (size_t j = 0; j<target.size() + 1; j++)
 	{
 		table[0][j] = j;
 	}
 
-	for (int i = 1; i<source.size() + 1; i++) 
+	for (size_t i = 1; i<source.size() + 1; i++)
 	{
-		for (int j = 1; j<target.size() + 1; j++) 
+		for (size_t j = 1; j<target.size() + 1; j++)
 		{
 			if (source.at(i - 1) == target.at(j - 1)) 
 			{
